+++ ensayos/models.py
@@ -280,9 +280,9 @@ class ResultadosLIE (models.Model):
     
     ensayo= models.ForeignKey("LIE", on_delete=models.CASCADE, verbose_name="Ensayo LIE")
     concentracion= models.IntegerField(verbose_name="Concentración")
-    peso= models.DecimalField(decimal_places=2, max_digits=5, verbose_name="Peso equivalente")
-    pex= models.DecimalField(decimal_places=2, max_digits=5, verbose_name="PEX")
-    pm= models.DecimalField(decimal_places=2, max_digits=5, verbose_name="PM")
+    peso= models.DecimalField(decimal_places=1, max_digits=5, verbose_name="Peso equivalente")
+    pex= models.DecimalField(decimal_places=1, max_digits=5, verbose_name="PEX")
+    pm= models.DecimalField(decimal_places=1, max_digits=5, verbose_name="PM")
     dpdt=  models.IntegerField(verbose_name="dPdT")
     resultado= models.CharField(max_length=300, choices=resultadosPosibles, verbose_name= "resultado")
     
@@ -455,7 +455,7 @@ class ResultadosPmax (models.Model):
     
     ensayo= models.ForeignKey("Pmax", on_delete=models.CASCADE, verbose_name="Ensayo Pmax")
     concentracion= models.IntegerField(verbose_name="Concentración")
-    peso= models.IntegerField(verbose_name="Peso equivalente")
+    peso=models.DecimalField(decimal_places=1, max_digits=7, verbose_name="Peso equivalente")
     serie= models.CharField(max_length=300, choices=seriesPosibles, verbose_name= "Serie")
     pm= models.DecimalField(decimal_places=1, max_digits=7, verbose_name="PM")
     dpdt= models.IntegerField(verbose_name="dP/dT")
@@ -517,9 +517,9 @@ class ResultadosCLO (models.Model):
     
     ensayo= models.ForeignKey("CLO", on_delete=models.CASCADE, verbose_name="Ensayo LIE")
     concentracion= models.IntegerField(verbose_name="Concentración")
-    peso= models.DecimalField(decimal_places=2, max_digits=5, verbose_name="Peso equivalente")
-    pex= models.DecimalField(decimal_places=2, max_digits=5, verbose_name="PEX")
-    pm= models.DecimalField(decimal_places=2, max_digits=5, verbose_name="PM")
+    peso= models.DecimalField(decimal_places=1, max_digits=5, verbose_name="Peso equivalente")
+    pex= models.DecimalField(decimal_places=1, max_digits=5, verbose_name="PEX")
+    pm= models.DecimalField(decimal_places=1, max_digits=5, verbose_name="PM")
     dpdt=  models.IntegerField(verbose_name="dPdT")
 
     oxigeno= models.IntegerField(verbose_name= "Oxígeno")
